% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_spotify.R
\name{get_spotify_token}
\alias{get_spotify_token}
\title{Retrieve oauth2 token from spotify.utils-namespaced environment}
\usage{
get_spotify_token()
}
\description{
Retrieves a token if it is previously stored, otherwise prompts user to get one.
}
\keyword{internal}
