% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularity.R
\name{get_track_popularity}
\alias{get_track_popularity}
\title{Get popularity of one or more tracks on Spotify}
\usage{
get_track_popularity(trackid, simplyoutput = FALSE,
  stoken = get_spotify_token())
}
\arguments{
\item{trackid}{a character vector with Spotify IDs of tracks}

\item{simplyoutput}{choose if the output is a completed information or a simple data.frame with three columns (trackid,name,popularity)}

\item{stoken}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\value{
a data.frame
}
\description{
Obtain a popularity of a track ID o a vector of track IDs
}
\examples{
#Get track popularity of Madonna
\dontrun{
get_track_popularity("5U1tMecqLfOkPDIUK9SVKa")
}

}
