% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_spotify.R
\name{spotify_auth}
\alias{spotify_auth}
\title{Get Spotify Access Token}
\usage{
spotify_auth(client_id = Sys.getenv("SPOTIFY_CLIENT_ID"),
  client_secret = Sys.getenv("SPOTIFY_CLIENT_SECRET"), cache = TRUE,
  rdstoken = NA)
}
\arguments{
\item{client_id}{Spotify Web API application key. Default value is a System Environment variable "SPOTIFY_CLIENT_ID"}

\item{client_secret}{Spotify Web API application secret. Default value is System Environment variable "SPOTIFY_CLIENT_SECRET"}

\item{cache}{By default your credentials are locally cached in a file called
\code{.httr-oauth}. Set to FALSE if you need to authenticate separately
each time.}

\item{rdstoken}{File path to stored RDS token. In server environments where
interactive OAuth is not possible, a token can be created on a desktop
client and used in production. See examples.}
}
\value{
A Token2.0 object, invisibly
}
\description{
This function creates a Spotify access token.
}
\examples{
\dontrun{
get_spotify_access_token()
}

}
